# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "AGENT STATS",
    "description": "AGENT STATS STATUSES AND CHANNEL DATA"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "any description..."
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/ami/summary": {
      "get": {
        "tags": [
          "AMI"
        ],
        "description": "queue summary",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "status of queues.",
            "schema": {
              "$ref": "#/definitions/Summary"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/ami/summary/{name}": {
      "get": {
        "tags": [
          "AMI"
        ],
        "description": "status of queues.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "status of queues.",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/ami/status/{name}": {
      "get": {
        "tags": [
          "AMI"
        ],
        "description": "status of queues.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "status of queues.",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/ami/logout-agent": {
      "post": {
        "tags": [
          "AMI"
        ],
        "description": "logs out agent.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AgentLogOut"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/ami/pause/": {
      "post": {
        "tags": [
          "AMI"
        ],
        "description": "status of queues.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Pause"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/ami/un-pause/": {
      "post": {
        "tags": [
          "AMI"
        ],
        "description": "status of queues.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UnPause"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/ami/un-pause-memeber/": {
      "post": {
        "tags": [
          "AMI"
        ],
        "description": "status of queues.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UnPauseMemeber"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/ami/pause-memeber/": {
      "post": {
        "tags": [
          "AMI"
        ],
        "description": "status of queues.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PauseMemeber"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/ami/call-originate/": {
      "post": {
        "tags": [
          "AMI"
        ],
        "description": "status of queues.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Originate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/ami/call-barge/": {
      "post": {
        "tags": [
          "AMI"
        ],
        "description": "status of queues.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Barge"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/ami/call-spy/": {
      "post": {
        "tags": [
          "AMI"
        ],
        "description": "status of queues.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Spy"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/ami/status": {
      "get": {
        "tags": [
          "AMI"
        ],
        "description": "status of queues.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "status of queues.",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/ami/channels/{name}": {
      "get": {
        "tags": [
          "AMI"
        ],
        "description": "all current channels running on machine.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Channel"
            },
            "description": "all current channels running on machine."
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/ami/channels": {
      "get": {
        "tags": [
          "AMI"
        ],
        "description": "all current channels running on machine.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Channel"
            },
            "description": "all current channels running on machine."
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/ami/sip-peers": {
      "get": {
        "tags": [
          "AMI"
        ],
        "description": "all sip peers.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "schema": {
              "$ref": "#/definitions/SipPeers"
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/ami/cel-events": {
      "get": {
        "tags": [
          "AMI"
        ],
        "description": "all sip peers.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "schema": {
              "$ref": "#/definitions/SipPeers"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/ami/cdr-events": {
      "get": {
        "tags": [
          "AMI"
        ],
        "description": "all sip peers.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "schema": {
              "$ref": "#/definitions/SipPeers"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/ami/bridge": {
      "post": {
        "tags": [
          "AMI"
        ],
        "description": "all sip peers.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "schema": {
              "$ref": "#/definitions/SipPeers"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/ami/redirect": {
      "post": {
        "tags": [
          "AMI"
        ],
        "description": "all sip peers.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "schema": {
              "$ref": "#/definitions/SipPeers"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "Summary": {
      "type": "object",
      "properties": {
        "response": {
          "type": "object",
          "properties": {
            "success": {
              "type": "boolean",
              "example": true
            },
            "data": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "event": {
                    "type": "string",
                    "example": "QueueSummary"
                  },
                  "queue": {
                    "type": "string",
                    "example": "sales"
                  },
                  "loggedin": {
                    "type": "string",
                    "example": "0"
                  },
                  "available": {
                    "type": "string",
                    "example": "0"
                  },
                  "callers": {
                    "type": "string",
                    "example": "0"
                  },
                  "holdtime": {
                    "type": "string",
                    "example": "0"
                  },
                  "talktime": {
                    "type": "string",
                    "example": "0"
                  },
                  "longestholdtime": {
                    "type": "string",
                    "example": "0"
                  },
                  "actionid": {
                    "type": "string",
                    "example": "1660287273579"
                  }
                }
              }
            }
          }
        }
      }
    },
    "Status": {
      "type": "object",
      "properties": {
        "response": {
          "type": "object",
          "properties": {
            "success": {
              "type": "boolean",
              "example": true
            },
            "data": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "event": {
                    "type": "string",
                    "example": "QueueMember"
                  },
                  "queue": {
                    "type": "string",
                    "example": "sales"
                  },
                  "name": {
                    "type": "string",
                    "example": "SIP/salman"
                  },
                  "location": {
                    "type": "string",
                    "example": "SIP/salman"
                  },
                  "stateinterface": {
                    "type": "string",
                    "example": "SIP/salman"
                  },
                  "membership": {
                    "type": "string",
                    "example": "dynamic"
                  },
                  "penalty": {
                    "type": "string",
                    "example": "0"
                  },
                  "callstaken": {
                    "type": "string",
                    "example": "0"
                  },
                  "lastcall": {
                    "type": "string",
                    "example": "0"
                  },
                  "lastpause": {
                    "type": "string",
                    "example": "0"
                  },
                  "incall": {
                    "type": "string",
                    "example": "0"
                  },
                  "status": {
                    "type": "string",
                    "example": "5"
                  },
                  "paused": {
                    "type": "string",
                    "example": "0"
                  },
                  "pausedreason": {
                    "type": "string",
                    "example": ""
                  },
                  "wrapuptime": {
                    "type": "string",
                    "example": "0"
                  },
                  "actionid": {
                    "type": "string",
                    "example": "1660286739498"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "Pause": {
      "type": "object",
      "properties": {
        "queues": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "queue": {
                "type": "string",
                "example": "tenent1"
              },
              "reason": {
                "type": "string",
                "example": "some reason"
              }
            }
          }
        }
      }
    },
    "UnPause": {
      "type": "object",
      "properties": {
        "queues": {
          "type": "array",
          "example": [
            "tenent1",
            "tenent2"
          ],
          "items": {
            "type": "string"
          }
        }
      }
    },
    "PauseMemeber": {
      "type": "object",
      "properties": {
        "queue": {
          "type": "string",
          "example": "tenent1"
        },
        "interface": {
          "type": "string",
          "example": "PJSIP/3000"
        },
        "reason": {
          "type": "string",
          "example": "some reason"
        }
      }
    },
    "UnPauseMemeber": {
      "type": "object",
      "properties": {
        "queue": {
          "type": "string",
          "example": "tenent1"
        },
        "interface": {
          "type": "string",
          "example": "PJSIP/3000"
        }
      }
    },
    "Originate": {
      "type": "object",
      "properties": {
        "trunk": {
          "type": "string",
          "example": "name of trunk"
        },
        "number": {
          "type": "string",
          "example": "777968656"
        },
        "context": {
          "type": "string",
          "example": "dafault"
        },
        "mask": {
          "type": "string",
          "example": "9898989"
        },
        "exten": {
          "type": "string",
          "example": "88856472"
        }
      }
    },
    "AgentLogOut": {
      "type": "object",
      "properties": {
        "interface": {
          "type": "string",
          "example": "pjsip/1001"
        }
      }
    },
    "Spy": {
      "type": "object",
      "properties": {
        "channel": {
          "type": "string",
          "example": "only channel prfix like 6002 for PJSIP/6002-00000001 "
        },
        "number": {
          "type": "string",
          "example": "777968656"
        },
        "context": {
          "type": "string",
          "example": "dafault"
        },
        "mask": {
          "type": "string",
          "example": "9898989"
        },
        "exten": {
          "type": "string",
          "example": "88856472"
        }
      }
    },
    "Barge": {
      "type": "object",
      "properties": {
        "channel": {
          "type": "string",
          "example": "only channel prfix like 6002 for PJSIP/6002-00000001 "
        },
        "number": {
          "type": "string",
          "example": "777968656"
        },
        "context": {
          "type": "string",
          "example": "dafault"
        },
        "mask": {
          "type": "string",
          "example": "9898989"
        },
        "exten": {
          "type": "string",
          "example": "88856472"
        }
      }
    },
...
  }
}